Added test for dynamic loading fixed aws email to aws_ses and moved it into the correct folder fixed aws_cloudfront to cloudfront.
No changes to user-facing names of modules.
