+++ modules/tests/test_aws_email_processor.py
@@ -4,7 +4,7 @@ import os
 import pytest
 from pytest_mock import MockFixture
 
-from modules.aws_email import AwsEmailProcessor
+from modules.aws_ses.aws_ses import AwsEmailProcessor
 from opta.exceptions import UserErrors
 from opta.layer import Layer
 
@@ -36,7 +36,7 @@ class TestAwsEmailProcessor:
         aws_email_module = layer.get_module("awsses", 8)
 
         mocked_sesv2 = mocker.Mock()
-        mocked_boto3 = mocker.patch("modules.aws_email.boto3")
+        mocked_boto3 = mocker.patch("modules.aws_ses.aws_ses.boto3")
         mocked_boto3.client.return_value = mocked_sesv2
         mocked_sesv2.get_account.return_value = {"ProductionAccessEnabled": True}
         aws_dns_modules = layer.get_module("awsdns", 8)
@@ -54,9 +54,9 @@ class TestAwsEmailProcessor:
         )
         aws_email_module = layer.get_module("awsses", 8)
 
-        patched_logger = mocker.patch("modules.aws_email.logger")
+        patched_logger = mocker.patch("modules.aws_ses.aws_ses.logger")
         mocked_sesv2 = mocker.Mock()
-        mocked_boto3 = mocker.patch("modules.aws_email.boto3")
+        mocked_boto3 = mocker.patch("modules.aws_ses.aws_ses.boto3")
         mocked_boto3.client.return_value = mocked_sesv2
         mocked_sesv2.get_account.return_value = {
             "ProductionAccessEnabled": False,
@@ -79,9 +79,9 @@ class TestAwsEmailProcessor:
         )
         aws_email_module = layer.get_module("awsses", 8)
 
-        patched_logger = mocker.patch("modules.aws_email.logger")
+        patched_logger = mocker.patch("modules.aws_ses.aws_ses.logger")
         mocked_sesv2 = mocker.Mock()
-        mocked_boto3 = mocker.patch("modules.aws_email.boto3")
+        mocked_boto3 = mocker.patch("modules.aws_ses.aws_ses.boto3")
         mocked_boto3.client.return_value = mocked_sesv2
         mocked_sesv2.get_account.return_value = {
             "ProductionAccessEnabled": False,
@@ -104,11 +104,11 @@ class TestAwsEmailProcessor:
         )
         aws_email_module = layer.get_module("awsses", 8)
 
-        patched_prompt = mocker.patch("modules.aws_email.prompt")
+        patched_prompt = mocker.patch("modules.aws_ses.aws_ses.prompt")
         patched_prompt.side_effect = ["www.blah.com", "hello, world", "blah@runx.dev"]
-        patched_logger = mocker.patch("modules.aws_email.logger")
+        patched_logger = mocker.patch("modules.aws_ses.aws_ses.logger")
         mocked_sesv2 = mocker.Mock()
-        mocked_boto3 = mocker.patch("modules.aws_email.boto3")
+        mocked_boto3 = mocker.patch("modules.aws_ses.aws_ses.boto3")
         mocked_boto3.client.return_value = mocked_sesv2
         mocked_sesv2.get_account.return_value = {
             "ProductionAccessEnabled": False,
